# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "포포춘(For Fortune) 운세 API",
    "version": "2.0.0",
    "description": "AI 기반 운세 상담 플랫폼 - 사주, 타로, 꿈해몽, 행운번호 등"
  },
  "servers": [
    {
      "url": "http://localhost:3350",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 액세스 토큰을 Bearer 형식으로 전달"
      }
    },
...
  },
  "paths": {
...
    "/api/auth/refresh": {
      "post": {
        "operationId": "refreshToken",
        "summary": "토큰 갱신",
        "description": "리프레시 토큰을 사용하여 새로운 액세스 토큰을 발급합니다.",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "토큰 갱신 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string",
                          "description": "새로운 JWT 액세스 토큰"
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "nickname": {
                              "type": "string"
                            }
                          }
                        },
                        "expiresIn": {
                          "type": "integer",
                          "description": "토큰 만료 시간 (초)"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "토큰이 갱신되었습니다."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "토큰 갱신 실패",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "string",
                      "enum": [
                        "MISSING_AUTHORIZATION_HEADER",
                        "EMPTY_ACCESS_TOKEN",
                        "TOKEN_EXPIRED",
                        "TOKEN_MALFORMED",
                        "TOKEN_SIGNATURE_INVALID",
                        "TOKEN_INVALID",
                        "USER_NOT_FOUND",
                        "REFRESH_TOKEN_ERROR"
                      ],
                      "description": "에러 코드"
                    },
                    "error": {
                      "type": "string",
                      "description": "에러 메시지"
                    },
                    "message": {
                      "type": "string",
                      "example": "다시 로그인해주세요."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/temp-register": {
      "post": {
        "operationId": "createTempAccount",
        "summary": "임시 계정 생성 (결제 심사용)",
        "description": "결제 심사를 위한 임시 ID/비밀번호 계정을 생성합니다.",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "nickname"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "test@example.com",
                    "description": "로그인용 이메일"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "password123",
                    "description": "로그인용 비밀번호"
                  },
                  "nickname": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 40,
                    "example": "테스터",
                    "description": "사용자 닉네임"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "계정 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "user_123"
                        },
                        "email": {
                          "type": "string",
                          "example": "test@example.com"
                        },
                        "nickname": {
                          "type": "string",
                          "example": "테스터"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "임시 계정이 생성되었습니다."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "필수 정보 누락"
          },
          "409": {
            "description": "이메일 중복"
          }
        }
      }
    },
    "/api/auth/temp-login": {
      "post": {
        "operationId": "loginWithPassword",
        "summary": "ID/비밀번호 로그인 (결제 심사용)",
        "description": "이메일과 비밀번호로 로그인합니다.",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "test@example.com",
                    "description": "로그인용 이메일"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123",
                    "description": "로그인용 비밀번호"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string",
                          "example": "eyJ0eXAiOiJKV1Q..."
                        },
                        "refreshToken": {
                          "type": "string",
                          "example": "eyJ0eXAiOiJKV1Q..."
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "example": "user_123"
                            },
                            "email": {
                              "type": "string",
                              "example": "test@example.com"
                            },
                            "nickname": {
                              "type": "string",
                              "example": "테스터"
                            },
                            "role": {
                              "type": "string",
                              "example": "USER"
                            }
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "로그인에 성공했습니다."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "필수 정보 누락"
          },
          "401": {
            "description": "인증 실패"
          }
        }
      }
    },
    "/api/auth/account": {
      "delete": {
        "operationId": "deleteAccount",
        "summary": "계정 탈퇴",
        "description": "사용자 계정을 완전히 삭제합니다.",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "계정 삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "계정이 삭제되었습니다."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
...
  }
}